==> scripts_archive/2.Variable_creation.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment

```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to data channel
```{r Read in file with path to data channel on teams}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables
Add the remove_duplicates function - used to deduplicate and remove NAs from IDs
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
Add the imp_check function - used to check variables for implausible values
```{r Read in functions}
source(file = paste0("functions/add_numeric.R"))
source(file = paste0("functions/remove_duplicates.R"))
source(file = paste0("functions/package_check.R"))
```

Load packages
```{r Load packages}
library(tidyverse)
library(summarytools)
library(gtsummary)
```

Set current date
```{r current date}
date <- Sys.Date()
```

Load data
```{r Load data}
dat <- read_rds(
  file = paste0(filepath_cleaned_data, "merged_dat.rds")
)
```

Overview variables
```{r Overview variables}
names(dat)
```

# Self-report and algorithm-derived ED diagnosis
```{r deriving self-report and algorithm-derived ED diagnosis}
# Anorexia nervosa
dat <- dat %>%
  mutate(
    anorexia_nervosa =
      case_when(
        ed.DSM5_AN_binary == "DSM5 AN" |
          mhd.an_numeric == 1 ~ "Anorexia nervosa"),
    
    bulimia_nervosa =
      case_when(
        ed.DSM5_BN_binary == "DSM-5 BN" |
          mhd.bn_numeric == 1 ~ "Bulimia nervosa"),
    
    binge_eating_disorder =
           case_when(
             ed.DSM5_BED_binary == "DSM-5 BED" |
               mhd.bed_numeric == 1 ~ "Binge-eating disorder")
    
    )
```




```{r Mixed presentation}
dat <- dat %>%
  mutate(
    mixed_eating_disorders =
      case_when(
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "AN & BN & BED",
        
        bulimia_nervosa == "Bulimia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "BN & BED",
        
        anorexia_nervosa == "Anorexia nervosa" &
        binge_eating_disorder == "Binge-eating disorder" ~ "AN & BED",
      
        anorexia_nervosa == "Anorexia nervosa" &
        bulimia_nervosa == "Bulimia nervosa" ~ "AN & BN",
       
        anorexia_nervosa == "Anorexia nervosa" ~ "AN",
        
        bulimia_nervosa == "Bulimia nervosa" ~ "BN",
        
        binge_eating_disorder == "Binge-eating disorder" ~ "BED"
      )
  )
```

```{r}
dat %>%
  freq(mixed_eating_disorders)
```



```{r}
dat %>% 
  dim
```



```{r demographics}
exclude_columns_from_pivoting <-
  c(
    "dem.sex",
    "demographics.what_gender_do_you_identify_with",
    "dem.do_you_identify_as_transgender",
    "dem.what_is_your_sexual_orientation",
    "dem.what_is_your_ethnic_origin",
    "dem.highest_education_finegrained",
    "dem.what_is_your_current_employment_status",
    "dem.what_is_your_current_maritalrelationship_status",
    "dem.pack_year",
    "mhd.eating_disorder_received_treatment"
  )


dat_demographics <- dat %>%
  select(
    mixed_eating_disorders,
    dem.sex,
    demographics.what_gender_do_you_identify_with,
    dem.do_you_identify_as_transgender,
    dem.what_is_your_sexual_orientation,
    dem.what_is_your_ethnic_origin,
    dem.highest_education_finegrained,
    dem.what_is_your_current_employment_status,
    dem.what_is_your_current_maritalrelationship_status,
    dem.pack_year,
    mhd.eating_disorder_received_treatment
  )



dat_demographics_long <- dat %>%
  select(
    
  all_of(exclude_columns_from_pivoting),
  
    # Disorder
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder
    ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

dat_demographics_long %>%
  head()
```


# Table:  Age distribution of EDGI UK sample by age range and sex
```{r select age distribution data}
age_sex_distribution <- dat %>%
  select(
    mixed_eating_disorders,
         dem.dob_age, 
         dem.sex)

# Long format
exclude_columns_from_pivoting <-
  c(
    "dem.dob_age",
    "dem.sex"
    )

age_sex_distribution_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
         
        # Disorder
        anorexia_nervosa,
        bulimia_nervosa,
        binge_eating_disorder
        ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

age_sex_distribution_long
```

# Table: Self-report and algorithm derived eating disorder diagnosis distribution in EDGI UK sample, split by sex
```{r select age distribution data}
selfreport_algo_sex <- dat %>%
  select(
         dem.sex,
         # Algorithm-derived
         ed.DSM5_AN_binary,
         ed.DSM5_AN_restricting_binary,
         ed.DSM5_AN_restricting_binary,
         ed.DSM5_BED_binary, 
         ed.DSM5_BN_binary,
         # Add purging, NES, and all atypical disorder (AN, BN, BED) algorithms
         # Self-report
         mhd.an_numeric,                                                    
         mhd.atypical_an_numeric,                                  
         mhd.bn_numeric,                                                
         mhd.bed_numeric,                                                
         mhd.atypical_bn_numeric,                                         
         mhd.atypical_bed_numeric,                                         
         mhd.purging_disorder_numeric,                                
         mhd.nighteating_syndrome_numeric,                                 
         mhd.pica_numeric,                                      
         mhd.avoidantrestrictive_food_intake_disorder_numeric,              
         mhd.rumination_disorder_numeric,                          
         mhd.feeding_eating_disorder_numeric,                              
         mhd.other_eating_disorder_numeric)
         
# Long format
exclude_columns_from_pivoting <-
  c(
    "dem.sex",
     "ed.DSM5_AN_binary",
     "ed.DSM5_AN_restricting_binary",
     "ed.DSM5_AN_restricting_binary",
     "ed.DSM5_BED_binary", 
     "ed.DSM5_BN_binary",
      # +++Add purging, NES, and all atypical disorder (AN, BN, BED) algorithms
      # Self-report
     "mhd.an_numeric",                                                    
     "mhd.atypical_an_numeric",                                  
     "mhd.bn_numeric",                                                
     "mhd.bed_numeric",                                                
     "mhd.atypical_bn_numeric",                                         
     "mhd.atypical_bed_numeric",                                         
     "mhd.purging_disorder_numeric",                                
     "mhd.nighteating_syndrome_numeric",                                 
     "mhd.pica_numeric",                                      
     "mhd.avoidantrestrictive_food_intake_disorder_numeric",              
     "mhd.rumination_disorder_numeric",                          
     "mhd.feeding_eating_disorder_numeric",                              
     "mhd.other_eating_disorder_numeric"
  )

selfreport_algo_sex_long <- dat %>%
  select(
    all_of(exclude_columns_from_pivoting),
        ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )

selfreport_algo_sex_long
```




```{r Psychiatric comorbidities}
exclude_columns_from_pivoting <-
  c(
    "mhd.depression_and_anxiety",
    "mhd.depressive_disorders",
    "mhd.anxiety_disorders",   
    "mhd.ptsd_numeric",
    
    "mhd.obsessive_compulsive_disorders",
    
    "mhd.personality_disorder_diagnosed_numeric",
    
    "mhd.bipolar_and_schizophrenia",
    "mhd.psychotic_disorders",
    "mhd.bipolar_disorder_numeric",
    
    "mhd.autism_spectrum_disorder",
    "mhd.addadhd_numeric"
  )


# mixed 
dat_psych_dx <- dat %>%
  select(
    mixed_eating_disorders,
    
    all_of(exclude_columns_from_pivoting)
  )

dat_psych_dx_longer <- dat %>%
  select(
    
    all_of(exclude_columns_from_pivoting),
    
    # Diagnosis
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder
  ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )


dat_psych_dx_longer %>%
  head()
```


```{r Somatic dx exlcude from pivoting}
exclude_columns_from_pivoting <-
  c(
    "dem.epilepsy_or_convulsions"                                        
 ,"dem.migraines"                                                      
 ,"dem.multiple_sclerosis"                                             
 ,"dem.parkinsons_disease"                                             
 ,"dem.severe_memory_loss"                                             
 ,"dem.neurological_none"                                              
 ,"dem.neurological_dont_know"                                         
 ,"dem.neurological_prefer_not_to_answer"                              
 ,"dem.hay_fever"                                                      
 ,"dem.drug_allergy"                                                   
 ,"dem.drug_allergy_txt"                                               
 ,"dem.food_allergy"                                                   
 ,"dem.food_allergy_txt"                                               
 ,"dem.other_allergy"                                                  
 ,"dem.other_allergy_txt"                                              
 ,"dem.osteoporosis"                                                   
 ,"dem.osteoarthritis"                                                 
 ,"dem.rheumatoid_arthritis"                                           
 ,"dem.other_arthritis"                                                
 ,"dem.other_arthritis_txt"                                            
 ,"dem.allergy_none"                                                   
 ,"dem.allergy_prefer_not_to_answer"                                   
 ,"dem.allergy_dont_know"                                              
 ,"dem.asthma"                                                         
 ,"dem.emphysema_or_chronic_bronchitis"                                
 ,"dem.heart_attack_or_angina"                                         
 ,"dem.high_blood_cholesterol"                                         
 ,"dem.high_blood_pressure"                                            
 ,"dem.atrial_fibrillation"                                            
 ,"dem.stroke"                                                         
 ,"dem.respiratory_none"                                               
 ,"dem.respiratory_prefer_not_to_answer"                               
 ,"dem.respiratory_dont_know"                                          
 ,"dem.crohns_disease"                                                 
 ,"dem.ulcerative_colitis"                                             
 ,"dem.coeliac_disease"                                                
 ,"dem.diabetes_type_1"                                                
 ,"dem.diabetes_type_2"                                                
 ,"dem.pain_due_to_diabetes"                                           
 ,"dem.pain_due_to_virus"                                              
 ,"dem.digestive_none"                                                 
 ,"dem.digestive_dont_know"                                            
 ,"dem.digestive_prefer_not_to_answer"                                 
 ,"dem.breast_cancer"                                                  
 ,"dem.lung_cancer"                                                    
 ,"dem.stomach_cancer"                                                 
 ,"dem.colon_cancer"                                                   
 ,"dem.uterus_cancer"                                                  
 ,"dem.prostate_cancer"
 ,"dem.cancer_none"
 ,"dem.cancer_dont_know"
 ,"dem.cancer_prefer_not_to_answer"
 ,"dem.psoriasis"
 ,"dem.vitiligo"
 ,"dem.eczema"
 ,"dem.thyroid_disease"
 ,"dem.thyroid_disease_txt"
 ,"dem.skin_none"
 ,"dem.skin_dont_know"
 ,"dem.skin_prefer_not_to_answer"
  )
```


```{r Somatic comorbidities}
# mixed 
dat_somatic_dx <- dat %>%
  select(
    mixed_eating_disorders,
    
    all_of(exclude_columns_from_pivoting)
  )

dat_somatic_dx_longer <- dat %>%
  select(
    
    all_of(exclude_columns_from_pivoting),
    
    # Diagnosis
    anorexia_nervosa,
    bulimia_nervosa,
    binge_eating_disorder
  ) %>%
  pivot_longer(
    -exclude_columns_from_pivoting,
#    names_to = "Old_variable",
    values_to = "Disorder"
    ) %>%
  select(
    -name
  )


dat_somatic_dx_longer %>%
  head()
```



```{r}
save.image(file = paste0(filepath_cleaned_data, "tables.Rdata"))
```
