==> scripts_archive/4.Age.Rmd <==
---
title: "Demographics"
author: "Christopher Huebel, Helena Davies, Dina Monssen"
date: "2022-08-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Clear Global Environment

```{r clear global environment}
rm(list = ls(all.names = TRUE)) 
```

Read in file with path to data channel
```{r Read in file with path to data channel on teams}
source(file = "../credentials/paths.R")
```

Add the add_numeric function - used to convert character variables into numeric variables
Add the remove_duplicates function - used to deduplicate and remove NAs from IDs
Add the sumscores function - used to generate sumscores
Add the package_check function - used to install and load dependencies
Add the imp_check function - used to check variables for implausible values
```{r Read in functions}
source(file = paste0("functions/add_numeric.R"))
source(file = paste0("functions/remove_duplicates.R"))
source(file = paste0("functions/package_check.R"))
```

Load packages

```{r Load packages}
library(tidyverse)
library(gtsummary)
```

Set current date

```{r current date}
date <- Sys.Date()
```

Load data

```{r Load data}
load(file = paste0(filepath_cleaned_data, "tables.Rdata"))
```


```{r}
names(age_sex_distribution)
```



```{r format table}
summary_table <-
  gtsummary::tbl_strata(
    data = dat_summary_table, 
    strata = ed.DSM5_AN_binary,
    .tbl_fun =
      ~ .x %>%
      tbl_summary(
        
        by =  ed.DSM5_BED_binary,
        
        label = list(
        dem.which_gender_do_you_identify_with ~
          "Gender",
        dem.what_is_your_sexual_orientation ~
          "Sexual Orientation",
        dem.do_you_identify_as_transgender ~
          "Transgender"
        ),
                
          type = list(dem.which_gender_do_you_identify_with ~ 
            "categorical",
          dem.what_is_your_sexual_orientation ~
            "categorical",
          dem.do_you_identify_as_transgender ~
            "categorical"),

          statistic = list(
                             
            all_categorical() ~ "{n} ({p}%)",
            
            all_continuous2() ~ c("{mean} ({sd})",
                                  "{median} ({p25}, {p75})",
                                  "{min}, {max}"
                                  )
                ), 
              
              missing = "always")  %>% 
  add_n()
  )
```

Add bold labels to table

```{r table with bold labels}
summary_table %>% 
  bold_labels() 
```

Save as a word document

```{r table as word document}
summary_table %>%
  as_flex_table() %>%
  flextable::save_as_docx(
    path = "../results/ARFID_ASD_sex.docx"
  )
```




Label variables for table and select type of data

```{r tbl_summary}
summary_table <- summary %>% 
  tbl_summary(
              label = list(
                child.parent.pre.biological_sex_of_your_child_at_birth ~ 
                  "Patient's biological sex",
                 child.parent.pre.age_yrs_dec_coalesced ~ 
                  "Age in years",
                 asd_diagnosis ~ 
                  "Comorbid autism spectrum disorder diagnosis",
                 pardi.parent.pre.current_bmi_age_sds ~ 
                  "BMI-for-age z-score",
                 pardi.parent.pre.diagnostic_predictor_a1 ~ 
                  "A1: Weight loss or difficulty maintaining weight/growth", 
                 pardi.parent.pre.diagnostic_predictor_a2 ~ 
                  "A2: Nutritional deficiency", 
                 pardi.parent.pre.diagnostic_predictor_a3 ~ 
                  "A3: Enteral feeding or oral supplement", 
                 pardi.parent.pre.diagnostic_predictor_a4 ~ 
                  "A4: Psychosocial impairment"),
              
              
              type = list(
                child.parent.pre.biological_sex_of_your_child_at_birth ~ 
                  "categorical",
                 child.parent.pre.age_yrs_dec_coalesced ~ 
                  "continuous2",
                 asd_diagnosis ~
                  "categorical",
                 pardi.parent.pre.current_bmi_age_sds ~ 
                  "continuous2",
                 pardi.parent.pre.diagnostic_predictor_a1 ~ 
                  "categorical", 
                 pardi.parent.pre.diagnostic_predictor_a2 ~ 
                  "categorical", 
                 pardi.parent.pre.diagnostic_predictor_a3 ~ 
                  "categorical", 
                 pardi.parent.pre.diagnostic_predictor_a4 ~ 
                  "categorical"), 
              
              statistic = list(
                all_categorical() ~ "{n} ({p}%)",
                all_continuous2() ~ c("{mean} ({sd})",
                                      "{median} ({p25}, {p75})",
                                      "{min}, {max}")
                ), 
              
              missing = "no"
    ) 
```

```{r modify labls and run table}
summary_table %>% 
  modify_caption("**Table 1. Patient Characteristics**") %>%  
  modify_footnote(all_stat_cols() ~ "n(%) or Mean(SD)") %>% 
  bold_labels() %>% 
  add_n() %>% 
  modify_header(label ~ "**Variable**") 
```
